Add cedilla to UK intl. with dead keys
AltGr+, used to map to comma, which is already available by simply
typing comma. Before this change, there was no way to output a cedilla
from the UK intl. keyboard.

The cedilla is important for Portuguese (one of the top 10 biggest
languages in the world). So much so, that the US intl.  already maps
AltGr+, to cedilla. This patch implements the same feature in the UK
intl. keyboad.